pyproject.toml
[project]
name = "app_1"
version = "0.1.0"
description = ""
authors = [{name = "Manuel Liébana Cazalla", email = "manu.lc46@gmail.com"}]
readme = "README.md"
requires-python = ">=3.11"      
dependencies = [
    "langchain>=0.2.11,<0.3",
    "langchain-openai>=0.1.20,<0.2",
    "python-dotenv>=1.0.1,<2",
    "jupyterlab>=4.2.4,<5",
    "langchain-community>=0.2.10,<0.3",
    "langchain-ollama>=0.1.3",
    "langchain-core>=0.2.43",
]

[build-system]
requires = ["hatchling"]   # uv se lleva bien con hatchling
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app_1"]
